Move graphql stuff around
